+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Configurable WebFinger output path with `activitypub.webfinger_output_file`.
+
 ## [0.9.0] - 2026-09-04
 
 ### Changed